backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
aiomysql==0.2.0
aiosqlite==0.20.0
PyMySQL==1.1.1
pydantic-settings==2.7.0
eval-type-backport==0.2.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
langchain==0.3.14
langchain-openai==0.2.14
langgraph==0.2.62
httpx==0.28.1
PyYAML==6.0.2
pytest==8.3.4
pytest-asyncio==0.25.0
asgi-lifespan==2.1.0
